requirements.txt
einops
matplotlib
mkl_service
numpy
pandas
PyYAML
scikit_learn
scipy
tensorflow
torch
tqdm